Pushing texel format translation into header
